main.jsの記述をTodoクラスへ切り出し
